fix(actions): publish chart independently
The Helm chart will be modified and released independently from tagged
versions of the controller. This allows us to update and test the chart
after new releases of the controller, as well as making improvements or
bugfixes to it.

This changeset modifies the action to publish new versions of charts
whenever there is a change on the main branch.